benchmarks/requirements.txt
asn1crypto==1.2.0
atomicwrites==1.3.0
attrs==19.3.0
bcrypt==3.1.7
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
docker==3.7.0
docker-pycreds==0.4.0
idna==2.8
importlib-metadata==0.23
more-itertools==7.2.0
packaging==19.2
paramiko==2.6.0
pathlib2==2.3.5
pexpect==4.7.0
pluggy==0.9.0
ptyprocess==0.6.0
py==1.8.0
pycparser==2.19
PyNaCl==1.3.0
pyparsing==2.4.5
pytest==4.3.0
PyYAML==5.1.2
requests==2.22.0
six==1.13.0
urllib3==1.25.7
wcwidth==0.1.7
websocket-client==0.56.0
zipp==0.6.0
